modules/rtv: fix settings path for channel settings
The channel module name was being used as the settings instance
name, which is redundant to the entries[0] "value as label"
behavior with the module name also @ entries[0].

The resulting paths resembled:

/module/rtv/compose/compose/layers/[N]/...
/module/rtv/compose/compose/texture/...

Now they are:

/module/rtv/channel/compose/layers/[N]/...
/module/rtv/channel/compose/texture/...

There's still work to be done in this area of rtv.  It's unclear
if even a static "/module/rtv/channel" is correct, or if it
should show the subscript of the channels[] entry currently being
used when that becomes a proper nested settings situation, i.e.

/module/rtv/channels/[0]/compose/layers/[N]/...
...
/module/rtv/channels/[1]/spiro
...
/module/rtv/channels/[2]/roto
...

for a "channels=compose,spiro,roto" kind of hypothetical

another option which might make sense is to have the channel path
more like an auto-increment identifier so the channel contexts
have unique paths and don't collide on the stream.  This could be
important for scenarios utilizing the "ref" module and trying to
reuse parts of the context from one channel to the next.  When
the paths collide things tend to break in weird ways presently.

this area needs more thought