modules/roto: implement fill_module= setting
This makes it possible to tiled+rotate the output of another
module in the same manner checkers::fill_module fills cells with
module output.

The default stays "none" for the classic roto with the
psychedelic color cycling.  When !"none" the color cycling
doesn't get applied currently.  It might be interesting to try
support that in the future though.