main: fix UAF bug on shutdown
When introducing the **fragment_ptr model in 5a0776f, the
rototiller_thread() introduced a local place to put the pointer
to point at when rendering.

But this pointer then ends up outliving the thread on shutdown
within any queued frames until quiescent.  Fixed in the obvious
way by sticking it in rototiller_t instead.