modules/shapes: drop atan2_approx()
With the addition of the "radcache" in b6362c, the need for a
faster approximate atan2f() is largely eliminated.

And there seems to be a bug in the implementation as-is taken
from https://mazzo.li/posts/vectorized-atan2.html

You can see the bug as vertical line artifact around the center
where the X coordinate would be 0.  Rather than debug what's
wrong with this approximation's implementation surrounding its
quadrant adjustments, let's just resume using atan2f() and let
the cache keep things quick.