modules/rkt: s/_EDITSCENE/_SCENE/
This was a bit of an aspirational misnomer, editing scenes isn't
actually implemented yet.  What the EDITSCENE state currently
implements is the per-scene dialog+prompt, which originally was
going to just be the scene editing flow but became more of a
"view a specified scene's details" with a prompt of its own.

Nothing functionally changes, just mechanical internal renames.