til: introduce til_module_render_limited() variant
With the introduction of discoverable on-stream contexts, with
the intention of modules finding them at runtime for nested
rendering, relying exclusively on limiting n_cpus @
til_module_create_context() is no longer adequate.

When the nested rendering makes use of a context created
elsewhere, it can't make any assumptions about what its n_cpus
is, nor should it be attempting to change that value for its use.

So this variant just adds a max_cpus parameter for setting an
upper bound to whatever is found in the context's n_cpus.

The impetus for this is to fix the ref builtin, which you can
currently trick into deadlock by performing a nested threaded
render within a threaded render.  What the ref module should be
doing is propagating its own context's n_cpus down as the upper
bound, via this new render variant.