til: add til_fragmenter_slice_per_cpu_x16() variant
This restores the original til_fragmenter_slice_per_cpu() while
adding an explicit x16 variant for what
til_fragmenter_slice_per_cpu() had become.

The impetus for this is realizing the x16 multiplier is terrible
for sparkler's crappy threading, and it really need strictly
n_cpus slices for the threading to be beneficial.

So I'm just getting rid of the hidden x16 in favor of making it
explicit in an _x16 variant.

Subsequent commit will pivot all the non-sparkler callers to
til_fragmenter_slice_per_cpu_x16().