modules/signals: experimental signals module
Playing with libs/sig in 2D, this isn't really an interesting
module by itself in terms of visual output.  But it might have
utility as a diagnostic thing if libs/sig becomes a more used
thing.

At the very least, for now, it's useful for observing affects of
and iterating on libs/sig development.  So I'm merging this, just
gated behind TIL_MODULE_EXPERIMENTAL so it's not in rtv rotation
or presented as something in the usual modules list.