modules/rkt: make "blank" default scenes=
The "compose,compose,compose,compose" default was never intended
to be permanent, but gave a set of scenes to test the Rocket
integrations like scene selection and scene-specific tracks
without any additional effort.

Now that there's scener for easily adding/editing scenes, and
things are just generally more mature, I think it makes sense to
just go back to something minimal here.

I'd really rather just have it be "", but that's not handled well
presently.  There isn't really a way to start with an empty
scenes set for rkt.  Which is awkward, but "blank" is close.

It'd just be nice to start with an altogether blank slate rather
than having to always edit the default first scene when starting
anew...