modules/rkt: make the default base= "rkt"
The "tiller" base (base being a Rocket concept) was always a bit
spurious.  Maybe "til" would make more sense, but "rkt" is more
contextually specific.

I think when I originally picked "tiller" I was prioritizing
picking something unlikely to collide with another directory
name.  But the way Rocket is naming the directory in the
filesystem it gets suffixed with an _ anyways.