modules/rkt: use the raw module name when randomizing
The existing code assumed til_module_t.name was equivalent to the
name setting's value.  That's no longer a safe assumption, and if
the module was made possible via something like nocheck, that
needs to be preserved in the randomized setup.

So this gets the name out of the setting instead, using the new
raw value getter.