til_builtins: add "none" builtin
Sometimes nested modules are optional, in those cases I've
usually been using the "none" module name to indicate this and
the ad-hoc setup stuff can easily bypass module setup on that.

This commit is a step towards having a "none" builtin that
provides a til_module_t.setup() which succeeds but only produces
a NULL res_setup when asked to finalize.

So if callers just handle a successful finalize case that writes
NULL at res_setup equivalent to "disabled", this all Just Works.

Currently that's not an expected thing, but future commits will
bring everything else on the same page.

For callers who want to require the module and not offer "none",
they can just put "none" in the explicit exclusions list passed
to til_module_setup_full().