modules/roto: only update fill_fb on changed ticks
In compositions where roto recurs in the same frame, it's
wasteful to keep rendering the fill_fb contents manifold for the
same tick.  They shouldn't be varying anyways...

It'd be different if the fill_module rendered directly to the
output fragment in an overlay manner, but the current
roto implementation doesn't do that.  It's just sampling fill_fb
like a tiled texture sampler would.

The performance hit here is easily observed by doing something
like:

--module='checkers,size=4,fill_module=roto\,fill_module\=moire' --video=mem --defaults --go

Before this change, FPS=1 on my i7 X230.  After, FPS=~400.