modules/mixer: %S/fade/blend/g
Blend is a more accurate word for what's being done.

Fading is something you can achieve with the blend mode, if you
use something like "blank" as one of the modules.  But when you
have two modules generating content, it's not a fade, it's just a
blend.

It might make sense in the future to support a "fade" which
assumes a solid color for one module rather than having to use
the "blank" builtin which burns memory bandwidth on a blank
frame.