til_stream: always assign n_module_contexts in register
Leaving the larger n_module_contexts on reuse creates the
potential for a NULL ptr dereference in places like gc which
check the refcount of each countext in the set, without testing
for NULL.

Rather than making everything check for NULL, just shrink the
count when it gets reused in a smaller case.  This whole reuse
thing is kind of a silly little optimization anyways, and will be
revisited when cloning either happens or the sets become
deprecated.

For now just prevent creating situations that can crash.