til_fb: introduce til_fb_halt()
This makes it so til_fb_page_get() will return NULL when the fb
has been "halted".  If til_fb_page_get() is already waiting for a
page to become available, it will be woken up by til_fb_halt().

Preparatory commit for eliminating the reliance on
pthread_cancel() to terminate rototiller_thread() on shutdown.

Cancellation is always a PITA, especially cross-platform, but in
general it's so easy to have subtle bugs when using it.