modules/rtv: don't gc from cleanup_channel()
cleanup_channel() is called from rtv_destroy_context(), which
can be called by a gc.  The gc isn't reentrant and generally
isn't expecting this, so it breaks in hilarious ways when this
happens, usually ending in a segfault.

gc should probably get some defenses for asserting on reentry...
for now let's just prevent crashing by removing the gc call from
the destroy path.