til_fb: return queue delay from til_fb_page_get()
til_fb is triple buffered, but when rendering lags behind the
queue is empty making present-to-delay intervals smaller than
when rendering is keeping up and the queue is kept full.

This returns the duration of the returned pages last
submit-to-present delay, effectively measuring how long a page
can expect to wait to be flipped to/presented currently.

A subsequent commit will apply this delay value to the ticks
supplied to rendering.