main: apply submit-to-present delay to ticks
This should push the ticks value ahead by 1-2 frames worth of
time, when rendering is meeting/exceeding the frame rate.  Which
is the appropriate thing to do, since rendering is effectively
slightly ahead of the clock, producing visuals for now+N-frames
into the future.

When rendering lags behind, there's basically no delay, and
rendering is just operating on the "now" ticks which will be
flipped to ASAP once submitted.

This will probably be revisited once audio is rolled in, since
that too will need to be kept in sync with the visuals'
perspective of time.