modules/strobe: s/period/hz/
Preparatory commit for exposing strobe::hz as a tap, it seems
awkward to work in periods especially in the track data.

Though I do like the 0-1 range of period, though that doesn't
even hold for slower than 1HZ frequencies so... it's kind of a
lie anyways.

At least if the track is called "hz" anyone will know what the
values mean and easily reason about them.  So I'm making the
setting consistent with the soon to be added "hz" tap.