modules/*: til_module_context_free() error paths
Several modules still had vestigial ad-hoc free() cleanups on
error paths in their create_context().

Largely mechanical change of replacing those with
til_module_context_free() which is more appropriate, since the
til_module_context_t holds a reference on the setup.  A plain
free() will leak that reference.

But it's only on create_context() failures which are uncommon,
so this was in practice mostly harmless...