modules/drizzle: handle baking errors in drizzle_setup()
More setup_func conversion to returning the failed setting on
errors during res_setup baking.

Also fixed a bug while here in the style_values error detection;
it was misusing nelems() on the array when it's NULL-terminated
with a sentinel.  But this was only triggered if a user force
overrided the setting with the :-prefix syntax, since otherwise
the setting had to be in the values set according to the
front-end.  It was known there'd prolly be bugs when adding that
: override prefix support.  A lot of the module-local setup
baking code has been neglected/bitrotted/carelessly changed over
time, depending on front-end values policing to keep things on
the rails.