modules/rkt: better handle EINVAL errors on finalize
Similar to setup_interactively(), rkt_scener needs to handle
EINVAL errors on res_setup baking @ finalize.

Until now it had handled EINVAL @ finalize by failing the
operation and returning to the main scenes prompt.

With this commit rkt_scener now returns the user to the failed
setting, enabling correcting the problem.

It's a little janky, but not too bad.  See comments for why.