til_setup: add another helper for setup_func baking fails
This introduces til_setup_free_with_ret_err()
which just does the common idiom of:

- free the setup
- return with err code

so all these failure cases can be reduced to just a direct return
of calling this function.

Simpler version of til_setup_free_with_failed_setting_ret_err(),
which could have been reuesed for this by making the settings-related
parameters optional... but this way the call sites are less verbose
and these are tiny helpers it's harmless.