modules/pixbounce: make pixbounce threaded
Pixbounce isn't a particularly costly thing to render, but when
used as part of a composition, any time wasted with idle CPUs is
CPU time potentially stolen from other layers which could be
utilizing those CPUs.

So in this commit I've done a rather minimal conversion of the
pixbounce code to support threaded rendering.  It basically
doubles+ lone pixbounce FPS in --video=mem tests here.