drm_fb: better handle empty modes/connectors results
Phil tripped over this when his TV powered off in the midst of
playing with rototiller over hdmi output.

The NULL mode/connector was being supplied as the
til_setting_spec_t.preferred value throwing an assert in
til_setting_desc_new().

Just detect this exception and return an error.