modules/flow: don't reap elements on screen-space exit
This is too aggressive and produces some undesirable visible
artifacts on the periphery, especially for slow-moving
small-size fields.

In such scenarios the elements near the edges would be
excessively pruned when the direction wandered off-screen, then
leaving an overly sparse region when the direction inevitably
wandered back.

This is still an issue but it's far less prominent when only
clipping to the flow field boundaries... since the FOV doesn't
quite encompass the edges of the flow field.  Now the elements
can survive wandering a bit off-screen, and re-enter.