til_utils: switch til_get_ncpus() to sysconf
The ad-hoc sys-based probe of cpus works fine on Linux, but it's
not really preferable when Linux's sysconf supports
_SC_NPROCESSORS_ONLN, and there's a chance non-Linux's will
support it too.

Supposedly even Emscripten supports this, and will report the
number of available WebWorkers through this interface, even with
pthreads emulation.  I'm curious if that actually works, getting
wasm builds of rototiller demos could be fun.