modules/asc: turn into a "built-in" module
This deviates from the established pattern of all built-in
modules being in til_builtins.c, but that's ok.

The main impetus for this is to just get asc out of rotation for
randomized compose layers, since it's so uninteresting.  But it's
also valid as a built-in since it's primarily for diagnostic
purposes anyways.