til: introduce til_module_t.render_audio entrypoint
Simple audio rendering modules need a way to render their audio
without affecting the fragment's state.  By providing a
til_module_t.render_audio() a module indicates it's only an audio
renderer, and its render won't trigger an implicit clear of the
frame's fragment.

There might be a new flag added to indicate when modules are
audio modules, but this is good enough for now.