til_stream: add til_stream_render()
This introduces a concept of a stream's module context, which is
used to render the stream via til_stream_render().

The context is set using til_stream_set_module_context(), but at
this time no reference is taken - it just puts the context into
the stream for use by the render calls.  The caller is
responsible for ensuring it persist as long as the stream uses
it.

This is a preparatory commit for implementing pre/post render
module contexts that get hooked into a stream.  The impetus for
this is supporting background music via modules like playit,
without requiring stuff like rkt scenes to always include the bgm
context in the scene's module context heirarchy.

Nothing is functionally different here, the commit only adds the
new interfaces for setting the context and rendering via stream
instead of til_module_render() of the context you'd now set on
the stream.