til: introduce til_module_t.render_proxy() method
Like til_module_t.render_audio(), this hooks into the render path
without triggering the implicit fragment clearing of the usual
render_fragment/prepare_frame/finish_frame methods.

But unlike render_audio(), it receives the full suite of
parameters including the fragment pointer.

The impetus for this is implementing builtins that serve as a
sort of pass-thru or proxy to another module, like the "ref"
builtin (which should probably change to use this instead of its
current render_fragment() hook).  Because sometimes the proxied
module isn't something drawing to the fragment, so the implicit
fragment clearing shouldn't occur.  But if the proxied module
/does/ touch the fragment by implementing one of those methods,
then the proxy module's use of til_module_render() would still
trigger the implicit clear when appropriate.