til_builtins/ref: s/render_fragment/render_proxy/
Since ref can be used to reference any module type by path,
including ones that don't touch the fragment, it's inappropriate
to integrate via render_fragment() since that will cause an
implicit clear of the fragment.

This is exactly the type of scenario render_proxy was added for,
so use it here.