til_builtins/pre: introduce "pre" pre-render hook module
This adds a new "meta" module which registers itself as a
pre-render context on the stream, to render whatever module is
specified as the "module=" setting.

The impetus for this is background music in rkt creations, so one
can avoid enduring the rigamarole of plugging something like a
"playit" instance into every scene.

With this "pre" builtin one can just use something like
"pre,module=playit,itfile=music.it" as a scene, and that will get
the playit module playing music.it hooked into the stream,
without the need to always have it present in every scene's
settings.

There's a TODO item noting the need for taps to control the
active/inactive state of the pre instance this adds.  For now,
once you have a "pre", whatever module it contains will be
rendered before the main stream context set
w/til_stream_set_module_context() gets rendered.  That's fine for
now, but there will need to be more control especially as more
complex productions happen involving multiple songs.