til,main: introduce ratio= --video setting
First stab at supporting explicit aspect ratios.

This performs the adjustment when needed in til_fb so it's
automatically applied to all fb backends.

The syntax is ratio=W:H with ratio=full being special cased for
when no aspect ratio adjustment is desired (just use whatever the
fb page dimensions are, usually specified via size= in the fb
backend's settings, or display native res when fullscreen=on)

For now when an aspect ratio is specified it always fits the
content within the alotted space... there is no
full-but-ratio-preserved-by-clipping-if-needed variant like widescreen
TVs often have.