modules/book: implement a flipbook module
This module cycles through a set of modules referred to as pages.

Settings include:

pages={list,of,modules}
rate=float page flipping rate
direction=float page flipping direction

taps are exposed for controlling rate and direction real-time,
with the addition of an explicit page index tap for overriding
the automatic flipping normally influenced by rate and direction.

Note that when using rkt with these taps, it immediately becomes
limiting that there's no ability to specify NaN as a track value
for temporarily relinquishing control of the page idx.  It'd be
very useful to let the automatic flipping run the show at
whatever rate the rkt tracks specify, until the page idx was
asserted to a specific page, then holding it on that page until a
NaN row for the page track was encountered, restoring automatic
flipping.  This needs to get added in both the Rocket library and
RocketEditor to be supported, as well as getting wired up to the
rkt integration.