modules/spokes: make spokes count configurable
This switches the perimeter and stride math to use floating
point, done in a kind of fast and nasty naive substitution
manner.

That's necessary for getting away from the even stride
calculation / spokes calculation in favor of supporting arbitrary
spoke counts, without introducing discontinuities at the
boundaries of the mirrored sides

The new setting is count= with a handful of reasonable presets.
It's a little awkward since this is expressed as half the actual
count, naming could probably be improved.