modules/mixer: introduce bottom= setting
This adds a bottom={a,b} setting when
style={interlace,paintroller,sine}

Note this only introduces the settings, not the implementation.

The impetus for this is when experimenting with mixer transitions
in a work-in-progress, you'll sometimes want to invert the bottom
vs. top layers just to see how that looks.  Instead of having to
transpose the a_module<->b_module setting values, this provides a
separate setting achieving that, which will simply be ignored for
the irrelevant mixer styles.

Conceptually "bottom" vs. "top" doesn't really apply to all the
styles however.  style=interlace already feels a bit forced to
include here, but in the code it's clearly part of a group with
paintroller and sine, so it's included here.

But one could argue that *all* styles should have some kind of
toggle to swap a_module<->b_module.  Except there will probably
be another setting introduced to invert T as well, which would
effectively achieve that for the modules where "bottom" vs. "top"
doesn't conceptually fit.